fixed browser back button and page load on bottom issues in app release process
