+++ RELEASENOTES.md
@@ -6,7 +6,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 ### 0.5.5
 
 * Bugfix - Connect Partner Network to BPDM
-
+* Bugfix - App Release Process - Fixed browser back button issue
+* Bugfix - App Release Process - Fixed user directing to bottom of the page
 
 ### 0.5.4
