+++ cx-portal/src/components/pages/AppReleaseProcess/AppReleaseProcess.scss
@@ -128,8 +128,4 @@
       }
     }
   }
-
-  .create-app-section {
-    margin-top: 112px;
-  }
 }
