+++ cx-portal/src/components/pages/AppReleaseProcess/components/ReleaseProcessSteps.scss
@@ -45,3 +45,12 @@
 .form-divider {
   margin: 64px -16px 16px -16px !important;
 }
+
+.create-app-section {
+  margin-top: 112px;
+}
+
+.container {
+  max-width: 1140px;
+  margin: 0 auto;
+}
